requirements.txt
# EchoTune AI - Python Dependencies

# Core ML and Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Spotify API Integration
spotipy>=2.22.0
requests>=2.31.0

# MCP Server Dependencies
aiohttp>=3.8.0
websockets>=11.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Database Support
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0
supabase>=2.3.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
tqdm>=4.65.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipython>=8.0.0

# Environment Management
python-dotenv>=1.0.0

# Audio Feature Analysis
librosa>=0.10.0  # For advanced audio analysis if needed

# Web Framework (for future API endpoints)
fastapi>=0.100.0
uvicorn>=0.22.0

# Logging and Monitoring
structlog>=23.0.0